+++ nest-forms-backend/src/nases/types/jwt-nases.types.ts
@@ -1,0 +1,23 @@
+class JwtNasesPayloadActor {
+  declare name: string
+
+  declare sub: string
+}
+
+export class JwtNasesPayload {
+  declare sub: string
+
+  declare exp: number
+
+  declare nbf: number
+
+  declare iat: number
+
+  declare name: string
+
+  declare actor: JwtNasesPayloadActor
+
+  declare scopes: string[]
+
+  declare jti: string
+}
